✨ 기능 등록
회전, 줌 진행중
