+++ Assets/Scripts/Command/RotateCommand.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c3e9782495652c34bbb3d0e8ae32fc05
