+++ Assets/Scripts/System/Command/CommandSystemManager.cs
@@ -38,44 +38,3 @@ public interface ICommand
     public void Execute(string output);
     public string GetInputFormat();
 }
-
-public class RotateCommand : ICommand
-{
-    private const string CommandName = "Rotate";
-    private const string MessageFormat = "\nplease convert this by only this json format. direction type : left, right, up, down \n";
-    private Transform targetTransform;
-
-    protected RotationInformation rotationInformation;
-    public RotateCommand(Transform target)
-    {
-        targetTransform = target;
-        CommandSystemManager.instance.RegisterCommand(CommandName, this);
-        rotationInformation = new RotationInformation();
-    }
-
-    public void Execute(string output)
-    {
-        try
-        {
-            rotationInformation = JsonUtility.FromJson<RotationInformation>(output);
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            throw;
-        }
-    }
-
-    public string GetInputFormat()
-    {
-        string message = MessageFormat + JsonUtility.ToJson(rotationInformation);
-        return message;
-    }
-}
-
-public class RotationInformation
-{
-    public float angle;
-    public string direction;
-}
-
